backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
boto3
psycopg2-binary